[datacenter] Make the request method datacenter aware [minor] Removed the `res.statusCode` code smell
Applications that are hosted in different datacenters need to be called at their
correct endpoint. The cloud details are inside of the app documents so we need to fetch
both the endpoint resource and the app document before we know which datacenter we
need to hit for the given change.

The statusCode was introduced in commit 9cea32719a797bbd769eaa89ea7cfab85c7f345d but
Avian has no memory of why he did it. So i'm removing it to prevent possible breakage
when you expect that the result is always an object.